Expose view state with MessageComposerViewModel
